+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix namespace handling when waiting for deployed apps.
+
 ## [0.8.0] - 2020-12-10
 
 ### Added